simplify imports and rename StatsProvider -> StatsLazyProvider
