+++ frontend/src/App.tsx
@@ -14,6 +14,7 @@ import DonatePage from './pages/about/Donate';
 import PersonalVouchersPage from './pages/personal/vouchers/VouchersPage';
 import Frame from './features/frame/Frame';
 
+import { StatsLazyProvider } from './features/comparisons/StatsContext';
 import PublicRoute from './features/login/PublicRoute';
 import PrivateRoute from './features/login/PrivateRoute';
 
@@ -30,7 +31,6 @@ import PollRoutes from './app/PollRoutes';
 import { PollProvider } from './hooks/useCurrentPoll';
 import FAQ from './pages/faq/FAQ';
 import { scrollToTop } from './utils/ui';
-import { StatsProvider } from './features/comparisons/StatsContext';
 
 // The Analysis Page uses recharts which is a rather big library,
 // thus we choose to load it lazily.
@@ -80,7 +80,7 @@ function App() {
 
   return (
     <PollProvider>
-      <StatsProvider>
+      <StatsLazyProvider>
         <ScrollToTop />
         <InitGlobalStates />
         <Frame>
@@ -148,7 +148,7 @@ function App() {
             ))}
           </Switch>
         </Frame>
-      </StatsProvider>
+      </StatsLazyProvider>
     </PollProvider>
   );
 }
