+++ frontend/src/hooks/index.ts
@@ -5,3 +5,4 @@ export { useLoginState } from './useLoginState';
 export { useListFilter } from './useListFilter';
 export { useNotifications } from './useNotifications';
 export { useRefreshSettings } from './useRefreshSettings';
+export { useStats } from './useStats';
